+++ database/migrations/2021_12_21_193335_create_invoice_items_table.php
@@ -17,7 +17,7 @@ class CreateInvoiceItemsTable extends Migration
         Schema::create('invoice_items', function(Blueprint $table) {
             $table->id();
             $table->smallInteger('quantity');
-            $table->bigInteger('price');
+            $table->decimal('price', 8, 2);
             $table->foreignIdFor(Item::class)->constrained();
             $table->timestamps();
         });
